Fixed native OnDecline callback not running when the app is in terminated state, when decine is clicked when there is an incoming call, it does nothing so i made this fixes
