+++ example/android/app/build.gradle
@@ -20,6 +20,9 @@ android {
         versionCode flutter.versionCode
         versionName flutter.versionName
         multiDexEnabled true
+
+        manifestPlaceholders["applicationName"] = "com.example.flutter_callkit_incoming_example.MainApplication"
+
     }
 
     compileOptions {
